# [Draft] :alembic: Set up django-cachalot