--- a/requirements/base.txt
+++ b/requirements/base.txt
@@ -81,6 +81,7 @@ django==3.2.18
     #   commonground-api-common
     #   django-admin-index
     #   django-axes
+    #   django-cachalot
     #   django-choices
     #   django-cors-headers
     #   django-db-logger
@@ -113,6 +114,8 @@ django-better-admin-arrayfield==1.4.2
     # via
     #   -r requirements/base.in
     #   mozilla-django-oidc-db
+django-cachalot==2.5.3
+    # via -r requirements/base.in
 django-choices==1.7.2
     # via
     #   -r requirements/base.in