--- a/src/openzaak/components/catalogi/tests/admin/test_admin_publish.py
+++ b/src/openzaak/components/catalogi/tests/admin/test_admin_publish.py
@@ -8,6 +8,7 @@ from django.urls import reverse
 from django.utils.translation import gettext as _
 
 import requests_mock
+from cachalot.api import cachalot_disabled
 from django_webtest import WebTest
 from freezegun import freeze_time
 
@@ -57,6 +58,7 @@ class ZaaktypeAdminTests(
     @tag("notifications")
     @override_settings(NOTIFICATIONS_DISABLED=False)
     @freeze_time("2022-01-01")
+    @cachalot_disabled()
     @patch("notifications_api_common.viewsets.send_notification.delay")
     def test_publish_zaaktype(self, m, mock_notif):
         procestype_url = (