Related issue: https://github.com/open-zaak/open-zaak/issues/1344

Notes:
- [ ] we should take a close look at the limits of this library: https://django-cachalot.readthedocs.io/en/latest/limits.html#limits
- [ ] cachalot seems to interact with debug toolbar in a weird way, some endpoints are a lot slower
- [ ] incompatibility with freezegun https://github.com/noripyt/django-cachalot/issues/126
- [ ] (how) will this affect CMIS?
- [ ] I disabled caching for the `ZaakIdentificatie` model, because this made some tests fail, I'm not sure if there are other identificatie tables