@stevenbal 2023-04-25
Locally this initially showed slightly unexpected results, some endpoints sped up quite drastically and some got slower (ZaakObjecten for example). After some investigation, I ran the tests again without `debug_toolbar` and this led to improved performance across the board:

Without cachalot:

|Type|Name|Request Count|Failure Count|Median Response Time|Average Response Time|Min Response Time|Max Response Time|Average Content Size|Requests/s|Failures/s|50%|66%|75%|80%|90%|95%|98%|99%|99.9%|99.99%|100%|
|---|---|---|---|---|---|---|---|---|---|---|---|---|---|---|---|---|---|---|---|---|---|
|GET|/besluiten/api/v1/besluiten|10|0|160.0|166.34742419992108|155.571608000173|175.7263299987244|50710.0|0.08434185376276264|0.0|170|170|170|170|180|180|180|180|180|180|180|
|GET|/catalogi/api/v1/besluittypen|10|0|110.0|150.3075821998209|96.77563399964129|291.5385630003584|69143.0|0.08434185376276264|0.0|110|110|230|250|290|290|290|290|290|290|290|
|GET|/catalogi/api/v1/informatieobjecttypen|10|0|97|98.03328349989897|93.38741999999911|106.4617600004567|35543.0|0.08434185376276264|0.0|98|98|99|100|110|110|110|110|110|110|110|
|GET|/catalogi/api/v1/resultaattypen|10|0|120.0|130.5869314001029|111.67400599879329|247.99031499969715|57302.0|0.08434185376276264|0.0|120|120|120|130|250|250|250|250|250|250|250|
|GET|/catalogi/api/v1/roltypen|10|0|110.0|120.49964050002018|99.78970500014839|247.26334800106997|26514.0|0.08434185376276264|0.0|110|110|110|110|250|250|250|250|250|250|250|
|GET|/catalogi/api/v1/statustypen|10|0|130.0|167.59685820034065|115.04213100124616|367.2993479995057|31373.0|0.08434185376276264|0.0|140|140|150|280|370|370|370|370|370|370|370|
|GET|/catalogi/api/v1/zaaktypen|10|0|590.0|607.1370884998032|509.68848999946204|789.02677699989|212013.0|0.08434185376276264|0.0|590|660|660|680|790|790|790|790|790|790|790|
|GET|/documenten/api/v1/enkelvoudiginformatieobjecten|10|0|520.0|516.6858902002787|499.34752700028184|531.7125030014722|135621.0|0.08434185376276264|0.0|520|520|520|520|530|530|530|530|530|530|530|
|GET|/zaken/api/v1/resultaten|10|0|160.0|162.74311270026374|120.00395999893954|179.65429200012295|35207.0|0.08434185376276264|0.0|170|170|170|170|180|180|180|180|180|180|180|
|GET|/zaken/api/v1/rollen|10|0|190.0|195.62945769976068|145.1978719996987|231.8549269984942|58202.0|0.08434185376276264|0.0|200|210|210|220|230|230|230|230|230|230|230|
|GET|/zaken/api/v1/statussen|10|0|2800.0|2826.9730176998564|2763.6931919987546|2901.1806010003056|40006.0|0.08434185376276264|0.0|2800|2800|2900|2900|2900|2900|2900|2900|2900|2900|2900|
|GET|/zaken/api/v1/zaakobjecten|10|0|330.0|326.73650619944965|284.1468379992875|363.71807899922715|39576.0|0.08434185376276264|0.0|330|330|340|340|360|360|360|360|360|360|360|
|GET|/zaken/api/v1/zaken|10|0|6300.0|6385.163565600124|6230.295441000635|6664.979002998734|141191.0|0.08434185376276264|0.0|6300|6400|6500|6600|6700|6700|6700|6700|6700|6700|6700|
||Aggregated|130|0|180.0|911.8800275845878|93.38741999999911|6664.979002998734|71723.15384615384|1.0964440989159143|0.0|190|330|520|590|2900|6300|6500|6600|6700|6700|6700|

With cachalot:

|Type|Name|Request Count|Failure Count|Median Response Time|Average Response Time|Min Response Time|Max Response Time|Average Content Size|Requests/s|Failures/s|50%|66%|75%|80%|90%|95%|98%|99%|99.9%|99.99%|100%|
|---|---|---|---|---|---|---|---|---|---|---|---|---|---|---|---|---|---|---|---|---|---|
|GET|/besluiten/api/v1/besluiten|59|0|120.0|132.06361203394746|115.35274500056403|276.37145900007454|50710.0|0.49311126391086774|0.0|120|120|130|130|130|260|270|280|280|280|280|
|GET|/catalogi/api/v1/besluittypen|60|0|88|106.85840134998823|79.6032220005145|255.8820280009968|69143.0|0.5014690819432553|0.0|88|91|93|98|220|240|250|260|260|260|260|
|GET|/catalogi/api/v1/informatieobjecttypen|59|0|91|90.7267257120481|82.4887819999276|100.72986100021808|35543.0|0.49311126391086774|0.0|91|93|94|95|96|98|98|100|100|100|100|
|GET|/catalogi/api/v1/resultaattypen|60|0|100.0|105.2241621331935|91.93876100107445|239.90762399989762|57302.0|0.5014690819432553|0.0|100|100|100|110|110|120|130|240|240|240|240|
|GET|/catalogi/api/v1/roltypen|60|0|95|94.33276900017518|87.73838800152589|107.42532999938703|26514.0|0.5014690819432553|0.0|95|96|96|96|100|100|100|110|110|110|110|
|GET|/catalogi/api/v1/statustypen|60|0|110.0|125.45484296664047|101.98538400072721|288.1015299990395|31373.0|0.5014690819432553|0.0|110|110|120|120|140|270|280|290|290|290|290|
|GET|/catalogi/api/v1/zaaktypen|60|0|240.0|275.62935169999037|219.0121679996082|672.1551549999276|212013.0|0.5014690819432553|0.0|240|250|260|380|400|400|430|670|670|670|670|
|GET|/documenten/api/v1/enkelvoudiginformatieobjecten|59|0|120.0|145.63650372881472|112.7024680008617|523.6023850011406|135621.0|0.49311126391086774|0.0|120|130|130|130|270|280|280|520|520|520|520|
|GET|/zaken/api/v1/resultaten|60|0|110.0|120.02509811673008|103.82883199963544|255.8640759998525|35207.0|0.5014690819432553|0.0|110|120|120|120|130|190|260|260|260|260|260|
|GET|/zaken/api/v1/rollen|60|0|120.0|134.1529224499633|115.81902100078878|296.08661100064637|58202.0|0.5014690819432553|0.0|120|130|130|130|140|260|270|300|300|300|300|
|GET|/zaken/api/v1/statussen|60|0|110.0|154.39219168332178|100.99970900046173|2742.8224769992084|40006.0|0.5014690819432553|0.0|110|110|110|110|120|120|120|2700|2700|2700|2700|
|GET|/zaken/api/v1/zaakobjecten|60|0|130.0|131.59442711666998|115.6827049999265|387.98307800061593|39576.0|0.5014690819432553|0.0|130|130|130|130|140|140|150|390|390|390|390|
|GET|/zaken/api/v1/zaken|60|0|230.0|382.28043324985873|207.09610900121334|6659.976234999704|141191.0|0.5014690819432553|0.0|240|250|360|370|400|430|440|6700|6700|6700|6700|
||Aggregated|777|0|120.0|153.84023118790952|79.6032220005145|6659.976234999704|71714.5250965251|6.494024611165156|0.0|120|120|130|160|240|280|400|420|6700|6700|6700|