@stevenbal 2023-05-01
> Kunnen we dit minimaal sanity testen door 1x lijst op te vragen, zaak toevoegen, en nog een keer de lijst opvragen, en checken of het nieuwe item er in zit?

ik heb dit lokaal getest met de Besluiten list en daar werkt dit. Voor zaken staat caching nu uit, omdat ik zag dat het cachen van de `ZaakIdentificatie` tabel een aantal falende tests veroorzaakte