@joeribekker 2023-06-22
Refinement: Since there are some "strange" issues we're not doing this right now but if there is -the need for more speed- we will have another look at this approach.